Added args

+++ grade_repo.py
@@ -9,6 +9,8 @@ from shutil import copyfile
 parser = argparse.ArgumentParser()
 parser.add_argument( "test_cases" )
 parser.add_argument( "dir", nargs="+" )
+parser.add_argument( "--dry", default=False, action="store_true" )
+parser.add_argument( "-v", "--verbose", default=False, action="store_true")
 args = parser.parse_args()
 
 def loadYAML( filename ) :
@@ -22,58 +24,68 @@ def loadYAML( filename ) :
 args.dir.sort(key=lambda x: x.split(".")[1])
 
 for path_dir in args.dir:
-    returncode = subprocess.call("git -C {} pull".format(path_dir).split())
-    print(path_dir.replace("/", ":"))
+    pull_command = "git -C {} pull".format(path_dir)
+    if args.verbose or args.dry:
+        print(pull_command)
+    if not args.dry:
+        code_pull = subprocess.call(pull_command.split())
+
+    if not args.dry and code_pull is not 0:
+        continue
+
     test_cases = loadYAML( args.test_cases )
 
     subpackage = test_cases["subPackage"]
     for exercise in test_cases["exercises"]:
         name = exercise["className"]
         path = "{}/**/{}/{}.java".format(path_dir, "/".join(subpackage.split(".")), name)
-        source_file = next(iter(glob( path, recursive=True)), None)
+        source_file = next(iter(glob(path, recursive=True)), None)
 
         print(source_file)
         if not source_file :
             print("{}: Not found".format(name))
             continue
 
-        #TODO [2:]
-        package = "/".join(source_file.split("/")[2:]).replace(".java","")
+        package = source_file.split("src/")[1].replace(".java","")
 
         out_dir = "{}/out/".format(path_dir)
         if not os.path.isdir(out_dir):
             os.makedirs(out_dir, exist_ok=True)
              
         # Build
-        returncode = subprocess.call("docker run --rm -v {}/{}:/app -w /app -i java:alpine javac -d out/ src/{}.java".format(os.getcwd(), path_dir, package).split())
-        if returncode != 0 :
-            print("Error compiling: {}.java".format(name))
-            continue
+        compile_command = "docker run --rm -v {}/{}:/app -w /app -i java:alpine javac -d out/ src/{}.java".format(os.getcwd(), path_dir, package)
+        if args.verbose or args.dry:
+            print(compile_command)
+        if not args.dry:
+            returncode = subprocess.call(compile_command.split())
+            if returncode != 0 :
+                print("Error compiling: {}.java".format(name))
+                continue
 
-        subprocess.call(["cat", source_file])
-        print()
+            subprocess.call(["cat", source_file])
+            print()
 
-        for test in exercise["tests"] :
-            test_input = test["input"]
-            test_output = test["output"]
-            # Run
-            process = subprocess.Popen("docker run --rm -v {}/{}:/app -w /app -i java:alpine java {}".format(os.getcwd(), out_dir, package).split(), stdin=subprocess.PIPE, stdout=subprocess.PIPE)
-            output = process.communicate(input=test_input.encode('utf-8'))[0].decode("utf-8")
-            if process.returncode != 0:
-                status = "RUNTIME"
-            elif len(output) == 0 :
-                status = "EMPTY"
-            elif output[-1] != "\n" :
-                status = "NEW_LINE"
-            else :
-                output = output.strip()
-                if output == test_output :
-                    status = "PERFECT"
-                elif test_output in output :
-                    status = "PASSED"
+            for test in exercise["tests"] :
+                test_input = test["input"]
+                test_output = test["output"]
+                # Run
+                process = subprocess.Popen("docker run --rm -v {}/{}:/app -w /app -i java:alpine java {}".format(os.getcwd(), out_dir, package).split(), stdin=subprocess.PIPE, stdout=subprocess.PIPE)
+                output = process.communicate(input=test_input.encode('utf-8'))[0].decode("utf-8")
+                if process.returncode != 0:
+                    status = "RUNTIME"
+                elif len(output) == 0 :
+                    status = "EMPTY"
+                elif output[-1] != "\n" :
+                    status = "NEW_LINE"
                 else :
-                    status = "FAILED"
-            print("- {:15}: {:8} input: \"{}\" expected_output: \"{}\" output: \"{}\"".format(test["name"], status, \
-                test_input.replace("\n","\\n"), test_output.replace("\n","\\n"), output.replace("\n","\\n")))
-        print()
+                    output = output.strip()
+                    if output == test_output :
+                        status = "PERFECT"
+                    elif test_output in output :
+                        status = "PASSED"
+                    else :
+                        status = "FAILED"
+                print("- {:15}: {:8} input: \"{}\" expected_output: \"{}\" output: \"{}\"".format(test["name"], status, \
+                    test_input.replace("\n","\\n"), test_output.replace("\n","\\n"), output.replace("\n","\\n")))
+            print()
     print("=================================")
